workaround to suppress warning on char == unsigned char arch.

+++ sys/netinet6/icmp6.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: icmp6.c,v 1.31 2000/06/12 16:21:03 itojun Exp $	*/
+/*	$NetBSD: icmp6.c,v 1.32 2000/06/13 16:34:37 itojun Exp $	*/
 /*	$KAME: icmp6.c,v 1.113 2000/06/12 09:24:41 itojun Exp $	*/
 
 /*
@@ -1416,7 +1416,7 @@ ni6_dnsmatch(a, alen, b, blen)
 		if (a - a0 + 1 > alen || b - b0 + 1 > blen)
 			return 0;
 
-		if (a[0] < 0 || b[0] < 0)
+		if ((int)a[0] < 0 || (int)b[0] < 0)
 			return 0;
 		/* we don't support compression yet */
 		if (a[0] >= 64 || b[0] >= 64)
